refactor(header): extract marshaling and unmarshaling
Move complex header serialization, encoding, and MAC computation logic into
dedicated types (`Marshaler`, `Unmarshaler`, `SectionEncoder`, `Serializer`,
`MACProcessor`). This improves modularity and separation of concerns.
